Rockman X Command Mission Password

+++ tools/bash/sigterm-cpu95.sh
@@ -1,27 +1,16 @@
 #!/bin/bash
-[ $(which pidof) ] || exit
-[ $(which ps   ) ] || exit
+[ $(which ps) ] || exit
 
-core=$(cat /proc/cpuinfo | grep processor | wc -l)
-let MAXCPU=$core*90
-echo "MAXCPU = $MAXCPU"
-export MAXCPU
-
-function cpu90kill {
+echo "start $0"
+function cpu95kill {
 	local cpu=${1%.*}  # float to int
 	local mem=${2%.*}  # float to int
 	local pid=$3
 	local cmd=${4##*/}
 	#echo "CPU = $cpu , MEM = $mem , PID = $pid , CMD = $cmd"
 
-	# CPU MEM
-	#   0   0  skip
-	#   0 100  kill
-	# 100   0  kill
-	# 100 100  kill
-	if (( $cpu < $MAXCPU && $mem < 90 )); then
-		return
-	fi
+	# whitelist
+	[[ "$cmd" == 'ffmpeg' ]] && return
 
 	# [BUG] firefox-bin
 	#   comm              cmd
@@ -31,11 +20,22 @@ function cpu90kill {
 	#   WebExtensions     $path/firefox-bin -contentproc -childID 2 -isForBrowser ... tab
 	#   Isolated Web Co   $path/firefox-bin -contentproc -childID 3 -isForBrowser ... tab
 	#   Web Content       $path/firefox-bin -contentproc -childID 4 -isForBrowser ... tab
-	kill    -15 $pid
-	killall -15 $cmd
-	echo "[cpu90kill] $@"
+	#
+	# = kill both process (by pid) and parent (by command)
+	if (( $cpu > 95 )); then
+		kill    -15 $pid
+		killall -15 $cmd
+		echo "[cpu > 95] $@"
+		return
+	fi
+	if (( $mem > 95 )); then
+		kill    -15 $pid
+		killall -15 $cmd
+		echo "[mem > 95] $@"
+		return
+	fi
 }
-export -f cpu90kill
+export -f cpu95kill
 
 ps=(
 	ps
@@ -45,8 +45,8 @@ ps=(
 	--sort   %cpu
 )
 while [ '1' ]; do
-	cpu90kill  $(${ps[@]} | tail -1)
-	sleep 5
+	cpu95kill  $(${ps[@]} | tail -1)
+	sleep 30
 done
 
 <<'////'
